+++ apps/desktop/src/components/v3/stackTabs/StackTab.svelte
