+++ apps/desktop/src/components/v3/stackTabs/StackTabMenu.svelte
