+++ apps/desktop/src/components/v3/stackTabs/StackTabNew.svelte
@@ -1,0 +1,284 @@
+<script lang="ts">
+	import newStackSvg from './assets/new-stack.svg?raw';
+	import dependentBranchSvg from './assets/dependent-branch.svg?raw';
+	import RadioButton from '$components/RadioButton.svelte';
+	import { stackPath } from '$lib/routes/routes.svelte';
+	import { StackService } from '$lib/stacks/stackService.svelte';
+	import { getContext } from '@gitbutler/shared/context';
+	import Button from '@gitbutler/ui/Button.svelte';
+	import Modal from '@gitbutler/ui/Modal.svelte';
+	import Textbox from '@gitbutler/ui/Textbox.svelte';
+	import Link from '@gitbutler/ui/link/Link.svelte';
+	import { slugify } from '@gitbutler/ui/utils/string';
+	import { goto } from '$app/navigation';
+
+	type Props = {
+		projectId: string;
+		// Currently selected stack id.
+		stackId?: string;
+		overflow?: boolean;
+	};
+
+	const { projectId, stackId, overflow = false }: Props = $props();
+	const stackService = getContext(StackService);
+	const [createNewStack, stackCreation] = stackService.newStack();
+	const [createNewBranch, branchCreation] = stackService.newBranch();
+
+	let createRefModal = $state<ReturnType<typeof Modal>>();
+	let createRefName: string | undefined = $state();
+	let createRefType: 'stack' | 'dependent' = $state('stack');
+
+	const slugifiedRefName = $derived(createRefName && slugify(createRefName));
+	const generatedNameDiverges = $derived(!!createRefName && slugifiedRefName !== createRefName);
+
+	const firstBranchResult = $derived(
+		stackId ? stackService.branchAt(projectId, stackId, 0).current : undefined
+	);
+	const firstBranchName = $derived(firstBranchResult ? firstBranchResult.data?.name : undefined);
+
+	function handleOptionSelect(event: Event) {
+		const target = event.target as HTMLInputElement;
+		createRefType = target.id === 'new-stack' ? 'stack' : 'dependent';
+	}
+
+	async function addNew() {
+		if (createRefType === 'stack') {
+			const data = await createNewStack({
+				projectId,
+				branch: { name: createRefName }
+			});
+			goto(stackPath(projectId, data.id));
+			createRefModal?.close();
+		} else {
+			if (!stackId || !createRefName) {
+				// TODO: Add input validation.
+				return;
+			}
+			await createNewBranch({
+				projectId,
+				stackId,
+				request: { targetPatch: undefined, name: createRefName }
+			});
+			goto(stackPath(projectId, stackId));
+			createRefModal?.close();
+		}
+	}
+
+	const isAddingNew = $derived(stackCreation.current.isLoading || branchCreation.current.isLoading);
+
+	// TODO: it would be nice to remember the last selected option for the next time the modal is opened
+</script>
+
+<button
+	aria-label="new stack"
+	type="button"
+	class="new-stack-btn"
+	class:overflow
+	onclick={() => createRefModal?.show()}
+>
+	<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<path d="M0 10H20M10 0L10 20" stroke="currentColor" stroke-width="1.5" />
+	</svg>
+</button>
+
+<Modal bind:this={createRefModal} width={500}>
+	<div class="content-wrap">
+		<Textbox
+			label="New branch"
+			id="newRemoteName"
+			bind:value={createRefName}
+			autofocus
+			helperText={generatedNameDiverges ? `Will be created as '${slugifiedRefName}'` : undefined}
+		/>
+
+		<div class="options-wrap">
+			<!-- Option 1 -->
+			<label for="new-stack" class="radio-label" class:radio-selected={createRefType === 'stack'}>
+				<div class="radio-btn">
+					<RadioButton checked name="create-new" id="new-stack" onchange={handleOptionSelect} />
+				</div>
+
+				<h3 class="text-13 text-bold text-body radio-title">New stack</h3>
+				<p class="text-12 text-body radio-caption">
+					Create an independent branch<br />in a new stack.
+				</p>
+
+				<div class="radio-illustration">
+					{@html newStackSvg}
+				</div>
+			</label>
+			<!-- Option 2 -->
+			<label
+				for="new-dependent"
+				class="radio-label"
+				class:radio-selected={createRefType === 'dependent'}
+			>
+				<div class="radio-btn">
+					<RadioButton name="create-new" id="new-dependent" onchange={handleOptionSelect} />
+				</div>
+				<h3 class="text-13 text-bold text-body radio-title">Dependent branch</h3>
+				<p class="text-12 text-body radio-caption">
+					Create a branch that depends on<br />the branches in the current stack.
+				</p>
+
+				<div class="radio-illustration">
+					{@html dependentBranchSvg}
+				</div>
+			</label>
+		</div>
+
+		<span class="text-12 text-body radio-aditional-info"
+			>{createRefType === 'stack'
+				? '└ Stacks that are currently applied will remain in the workspace.'
+				: `└ The new branch will be added on top of \`${firstBranchName}\``}</span
+		>
+	</div>
+
+	{#snippet controls(close)}
+		<div class="footer">
+			<span class="text-12 text-body footer-text"
+				>See more: <Link
+					target="_blank"
+					rel="noreferrer"
+					href="https://docs.gitbutler.com/features/stacked-branches#comparison-to-virtual-branches"
+					>Stacked vs. Dependent</Link
+				></span
+			>
+
+			<div class="footer__controls">
+				<Button kind="outline" type="reset" onclick={close}>Cancel</Button>
+				<Button
+					style="pop"
+					type="submit"
+					onclick={addNew}
+					disabled={!createRefName}
+					loading={isAddingNew}
+				>
+					{#if createRefType === 'stack'}
+						Add new stack
+					{:else}
+						Add dependent branch
+					{/if}
+				</Button>
+			</div>
+		</div>
+	{/snippet}
+</Modal>
+
+<style lang="postcss">
+	.new-stack-btn {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		border: 1px solid var(--clr-border-2);
+		border-bottom: none;
+		border-radius: 0 var(--radius-ml) 0 0;
+		height: 100%;
+		padding: 0 20px;
+		background: var(--clr-stack-tab-inactive);
+		color: var(--clr-text-3);
+		transition:
+			color var(--transition-fast),
+			background var(--transition-fast);
+
+		&:hover {
+			color: var(--clr-text-2);
+			background: var(--clr-stack-tab-inactive-hover);
+		}
+	}
+
+	.content-wrap {
+		display: flex;
+		flex-direction: column;
+		gap: 16px;
+	}
+
+	.options-wrap {
+		display: flex;
+		gap: 8px;
+	}
+
+	/* RADIO */
+	.radio-label {
+		--btn-bg: var(--clr-btn-ntrl-outline-bg);
+		--opacity-btn-bg: 0;
+		--btn-border-clr: var(--clr-btn-ntrl-outline);
+		--btn-border-opacity: var(--opacity-btn-outline);
+		/* illustration */
+		--illustration-outline: var(--clr-border-2);
+		--illustration-text: var(--clr-text-3);
+		--illustration-accent-outline: var(--clr-text-3);
+		--illustration-accent-bg: var(--clr-bg-2);
+
+		position: relative;
+		display: flex;
+		flex: 1;
+		flex-direction: column;
+		padding: 14px 14px 0;
+		gap: 4px;
+		transition:
+			border-color var(--transition-fast),
+			background-color var(--transition-fast);
+
+		border-radius: var(--radius-m);
+		background: color-mix(
+			in srgb,
+			var(--btn-bg, transparent),
+			transparent calc((1 - var(--opacity-btn-bg, 1)) * 100%)
+		);
+		border: 1px solid
+			color-mix(
+				in srgb,
+				var(--btn-border-clr, transparent),
+				transparent calc((1 - var(--btn-border-opacity, 1)) * 100%)
+			);
+
+		&:not(.radio-selected):hover {
+			--opacity-btn-bg: 0.14;
+		}
+	}
+
+	.radio-btn {
+		position: absolute;
+		right: 12px;
+		top: 12px;
+		display: flex;
+	}
+
+	.radio-caption {
+		opacity: 0.7;
+	}
+
+	.radio-illustration {
+		display: flex;
+		align-items: flex-end;
+		margin-top: 20px;
+		height: 100%;
+	}
+
+	.radio-aditional-info {
+		color: var(--clr-text-2);
+	}
+
+	/* MODIFIERS */
+	.radio-selected {
+		--btn-bg: var(--clr-theme-pop-bg);
+		--opacity-btn-bg: 1;
+		--btn-border-clr: var(--clr-btn-pop-outline);
+		/* illustration */
+		--illustration-outline: var(--clr-text-3);
+		--illustration-text: var(--clr-text-2);
+		--illustration-accent-outline: var(--clr-theme-pop-element);
+		--illustration-accent-bg: var(--clr-theme-pop-bg);
+	}
+
+	/* FOOTER */
+	.footer {
+		display: flex;
+		justify-content: space-between;
+		align-items: center;
+		width: 100%;
+		gap: 16px;
+		color: var(--clr-text-2);
+	}
+</style>
