+++ apps/desktop/src/components/v3/stackTabs/StackTabs.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import ReduxResult from '$components/ReduxResult.svelte';
-	import StackTab from '$components/v3/StackTab.svelte';
-	import StackTabMenu from '$components/v3/StackTabMenu.svelte';
-	import StackTabNew from '$components/v3/StackTabNew.svelte';
+	import StackTab from '$components/v3/stackTabs/StackTab.svelte';
+	import StackTabMenu from './StackTabMenu.svelte';
+	import StackTabNew from './StackTabNew.svelte';
 	import { stackPath } from '$lib/routes/routes.svelte';
 	import { StackService } from '$lib/stacks/stackService.svelte';
 	import { inject } from '@gitbutler/shared/context';
